Make canvasses more uniform
