+++ src/lib/p5-components/components/CanvasUtils.ts
@@ -1,6 +1,7 @@
 import type p5 from "p5";
-import type { Writable } from "svelte/store";
+import { writable, type Writable } from "svelte/store";
 import { setCanvasContext, type DrawFn } from "./CanvasContext";
+import type { Vector2 } from "three";
 
 // Protocall for drawing functions
 export class FnToDraw {
@@ -37,12 +38,45 @@ function setRelative(draw: DrawFn): DrawFn {
     };
 }
 
+export class Draggable {
+    position: Writable<Vector2>;
+    color: string;
+    key: Symbol;
+    defaultPosition: Vector2;
+
+    constructor(defaultPosition: Vector2, color: string) {
+        this.position = writable(defaultPosition.clone());
+        this.defaultPosition = defaultPosition.clone()
+        this.color = color;
+        this.key = Symbol('draggable');
+    }
+
+    reset() {
+        this.position.set(this.defaultPosition);
+    }
+}
+
+function createDragStore() {
+    const { subscribe, set, update } = writable<Draggable[]>([]);
+
+    return {
+        subscribe,
+        add: (draggable: Draggable) => update((draggables) => [...draggables, draggable]),
+        remove: (key: Symbol) => update((draggables) => draggables.filter(draggable => draggable.key !== key)),
+        reset: () => update((draggables) => { draggables.forEach(draggable => draggable.reset()); return draggables; }),
+        destroy: () => set([])
+    };
+}
+
+export const draggables = createDragStore();
+
 export type ContextParams = {
     mouseX: Writable<number>;
     mouseY: Writable<number>;
     width: Writable<number>;
     height: Writable<number>;
     scale: Writable<number>;
+    draggables: typeof draggables;
 }
 
 export function setDefaultContext(fnsToDraw: FnToDraw[], params: ContextParams) {
@@ -69,5 +103,6 @@ export function setDefaultContext(fnsToDraw: FnToDraw[], params: ContextParams) 
         width: params.width,
         height: params.height,
         scale: params.scale,
+        draggables: params.draggables
     });
 }
